% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsoriX-dataset.R
\docType{data}
\name{AssignDataAlien}
\alias{AssignDataAlien}
\title{Simulated assignment dataset}
\format{A \var{dataframe} with x observations on 3 variables:
\tabular{rlll}{ [, 1] \tab animalID \tab (\var{Factor}) \tab Identification
of the animal\cr [, 2] \tab species \tab (\var{Factor}) \tab Animal species
name\cr [, 3] \tab tissue.value \tab (\var{numeric}) \tab Deuterium delta
value of the tissue\cr }}
\description{
This dataset contains simulated deuterium delta values. 
The data can be used as an example to perform assignments using the function \code{\link{isofind}}.
}
\examples{

head(AssignDataAlien)
str(AssignDataAlien)

}
\seealso{
\code{\link{isofind}} to perform assignments
}
\keyword{datasets}
