% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsoriX-datasets.R
\docType{data}
\name{AssignDataBat2}
\alias{AssignDataBat2}
\title{Assignment dataset for bat species}
\format{
A *dataframe* with 244 observations on 3 variables:
\tabular{rlll}{
[, 1] \tab sample_ID \tab (*factor*) \tab Identification of the animal\cr
[, 2] \tab species \tab (*factor*) \tab Animal species name\cr
[, 3] \tab sample_value \tab (*numeric*) \tab Hydrogen delta value of the tissue\cr }
}
\source{
data directly provided by the authors of the following publication
}
\description{
This dataset contains data from Voigt, Lehmann and Greif (2015). It contains
hydrogen delta values of fur keratin from bats captured in 2008, 2009 and
2013 from their roosting sites in Bulgaria. We only retained the bats of the
genus Myotis from the original study. The data can be used as an example to
perform assignments using the function [isofind].
}
\examples{

head(AssignDataBat2)
str(AssignDataBat2)

}
\references{
Voigt, C.C., Lehmann, D., Greif, S. (2015). Stable isotope
ratios of hydrogen separate mammals of aquatic and terrestrial food webs.
Methods in Ecology and Evolution 6(11).
}
\seealso{
[isofind] to perform assignments
}
\keyword{datasets}
