% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsoriX-dataset.R
\docType{data}
\name{CalibDataAlien}
\alias{CalibDataAlien}
\title{Simulated calibration dataset}
\format{A \var{dataframe} with x observations on 6 variables:
\tabular{rlll}{ [, 1] \tab siteID \tab (\var{Factor}) \tab Identification of
the sampling site\cr [, 2] \tab long \tab (\var{numeric}) \tab Longitude
coordinate [decimal degrees]\cr [, 3] \tab lat \tab (\var{numeric}) \tab
Latitude coordinate [decimal degrees]\cr [, 4] \tab elev \tab
(\var{numeric}) \tab Elevation asl [m]\cr [, 5] \tab animalID \tab
(\var{Factor}) \tab Identification of the sampled animal\cr [, 6] \tab tissue.value
\tab (\var{numeric}) \tab Deuterium delta value of the tissue\cr }}
\description{
This dataset contains simulated deuterium delta values for corresponding locations
based on an assumed linear relationship between the animal tissue value and the 
deuterium delta values in the environment.
The data can be used as an example to fit a calibration model using the
function \code{\link{calibfit}}.
}
\details{
Users who wish to use their own dataset for calibration should create a
\var{dataframe} of similar structure than this one. The columns should possess 
the same names as the ones described above. If the elevation is unknown at the 
sampling sites, elevation information can be extracted from a high resolution elevation
raster using the function \pkg{\link[raster]{extract}}. In this dataset, we
retrieved elevations from the Global Multi-resolution Terrain Elevation Data
2010.
}
\examples{

data(CalibDataAlien)
head(CalibDataAlien)
str(CalibDataAlien)

## The following example require to have downloaded
## a large elevation raster with the function getelev()
## and will therefore not run unless you type:
## example(CalibDataAlien, run.dontrun=TRUE)

\dontrun{
if(require(raster)){
    ### DELETE AND RECREATE ELEVATION DATA
    CalibDataAlien$elev <- NULL  ## we delete them

    ## we reconstruct the elevation data using an elevation raster:
    ## (see ?getelev for details on how to get the tif file)
    elevationrasterbig <- raster("gmted2010_30mn.tif")
    CalibDataAlien$elev <- extract(
        elevationrasterbig,
        cbind(CalibDataAlien$long, CalibDataAlien$lat))
    head(CalibDataAlien)
}

}


}
\seealso{
\code{\link{calibfit}} to fit a calibration model
}
\keyword{datasets}
