% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4classes.R
\docType{class}
\name{CALIBFIT-class}
\alias{CALIBFIT-class}
\title{Class CALIBFIT}
\description{
Class CALIBFIT
}
\section{Slots}{

\describe{
\item{\code{method}}{a character string indicating the method used for the calibration}

\item{\code{species_rand}}{a logical indicating whether the species random effect is included in the model}

\item{\code{site_rand}}{a logical indicating whether the site random effect is included in the model}

\item{\code{param}}{the fixed-effect estimates of the calibration function}

\item{\code{fixefCov}}{the covariance matrix of the fixed effects}

\item{\code{phi}}{the residual variance of the calibration fit}

\item{\code{calib_fit}}{the fitted calibration model (if applicable)}

\item{\code{iso_fit}}{the fitted calibration model (if applicable)}

\item{\code{data}}{the calibration data}

\item{\code{sp_points}}{a list of spatial points used for calibration}
}}

