% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibfit.R
\name{calibmultifit}
\alias{calibmultifit}
\title{Fit the calibration model when the isoscape has been fitted using isomultifit}
\usage{
calibmultifit(calib.data, isofit, weighting = NULL, verbose = interactive(),
  control.optim = list())
}
\arguments{
\item{calib.data}{A \var{dataframe} containing the calibration data (see note
below)}

\item{isofit}{The fitted isoscape created by \code{\link{isofit}}}

\item{weighting}{An optional RasterBrick containing the weights}

\item{verbose}{A \var{logical} indicating whether information about the
progress of the procedure should be displayed or not while the function is
running. By default verbose is \var{TRUE} if users use an interactive R
session and \var{FALSE} otherwise.}

\item{control.optim}{A \var{list} to pass information to the argument control
of the optim call (for advanced users only).}
}
\value{
This function returns a \var{list} of class \var{calibmultifit}
  containing the averaged fixed-effect estimates of the calibration function,
  the averaged covariance of the fixed effects, the averaged residual
  variance of the calibration fit, the fitted calibration models, the
  original calibration data set with additional information added during the
  fit, and the location of the calibration points as spatial points.
}
\description{
This function is the counterpart of \code{\link{calibfit}} for the objects
created with \code{\link{isomultifit}}. It fits the calibration function for
each strata (e.g. month) defined by \code{split.by} during the call to
\code{\link{isomultifit}} and then average the fits. The function can handle
weighting for the aggregation process and can thus be used to perform
assignment onto annual averages precipitation weighted isoscapes.
}
\note{
See \code{\link{CalibDataAlien}} to know which variables are needed to
  perform the calibration fit and their names.
}
\examples{

## The examples below will only be run if sufficient time is allowed
## You can change that by typing e.g. IsoriX.options(example_maxtime = XX)
## if you want to allow for examples taking up to ca. XX seconds to run
## (so don't write XX but put a number instead!)

if(IsoriX.getOption("example_maxtime") > 60) {

## We prepare the data:
GNIPDataDEmonthly <- prepdata(data = GNIPDataDE,
                              month = 1:6,
                              split.by = "month")
 
## We fit the models for Germany:
GermanMultiFit <- isomultifit(iso.data = GNIPDataDEmonthly)

## We fit the calibration model:
calibMulti <- calibmultifit(calib.data = CalibDataAlien[1:99, ],
                       isofit = GermanMultiFit,
                       weighting = PrecipBrickDE)

plot(calibMulti)

## We fit the calibration model (equal weights):
calibMulti2 <- calibmultifit(calib.data = CalibDataAlien[1:99, ],
                       isofit = GermanMultiFit)

plot(calibMulti2)
}

}
\seealso{
\code{\link{IsoriX}} for the complete workflow
}
\keyword{models}
\keyword{regression}
