% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsoriX-datasets.R
\docType{data}
\name{CountryBorders}
\alias{CountryBorders}
\title{Borders of world CountryBorders}
\format{
A \emph{SpatVector} object
}
\source{
This \emph{SpatVector} is derived from the package
\pkg{rnaturalearth}. Please refer to this other package for description and
sources of this dataset. See example for details on how we created the
dataset.
}
\description{
This dataset contains a polygon polygon SpatVector (from \pkg{terra}).
It can be used to draw the borders of world countries.
}
\examples{

plot(CountryBorders, border = "red", col = "darkgrey")

## How did we create this file?

## Uncomment the following to create the file as we did
# if (require(rnaturalearth) && require(terra)) {
#    CountryBorders <- rnaturalearth::ne_countries(scale = 'medium', returnclass = 'sf')
#    CountryBorders <- vect(CountryBorders[, 0])
#    #saveRDS(CountryBorders, file = "IsoriX/inst/extdata/CountryBorders.rds", compress = "xz")
# }

}
\seealso{
\itemize{
\item \code{\link{OceanMask}} for another polygon used to embellish the plots
}
}
\keyword{datasets}
