% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsoriX-dataset.R
\docType{data}
\name{CountryBorders_terra}
\alias{CountryBorders_terra}
\title{Borders of world CountryBorders (terra format)}
\format{
A \emph{SpatialPolygons}
}
\source{
This \emph{SpatialPolygons} is derived from the package
\pkg{rnaturalearth}. Please refer to this other package for description and
sources of this dataset. See example for details on how we created the
dataset.
}
\description{
This dataset contains a polygon shapefile that can be used to plot the
borders of CountryBorders.
}
\examples{

if (require(sp)) {
   plot(CountryBorders_terra, border="red", col="darkgrey")
}

## How did we create this file?

## Uncomment the following to create the file as we did
# if (require(rnaturalearth) && require(terra)) {
#    CountryBorders <- rnaturalearth::ne_countries(scale = 'medium', returnclass = 'sp')
#    CountryBorders <- CountryBorders[0] # empty object
#    CountryBorders_terra <- vect(CountryBorders)
# }

}
\seealso{
\itemize{
\item \link{CountryBorders_terra} for the same file stored as a \pkg{terra} object
\item \link{OceanMask} for another polygon shapefile used to embellish the plots
}
}
\keyword{datasets}
