% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoscape.R
\name{futureisoscape}
\alias{futureisoscape}
\title{Predicts the spatial distribution of isotopic values}
\usage{
futureisoscape(elevation.raster, isofit, verbose = interactive())
}
\description{
This function is an alternative implementation of isoscape().
It is not exported but may be put in use in a future version of IsoriX.
It does not compute the predictions into chunks.
}
