% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsoriX-dataset.R
\docType{data}
\name{GermanFit}
\alias{GermanFit}
\title{The fitted isoscape model for Germany}
\format{An object of class \code{isofit}}
\description{
This dataset contains an object of class \code{isofit} containing the mean
model and residual dispersion model fitted on the data for Germany (see
example).
}
\details{
This fitted isoscape model has been obtained by running the function
\code{\link{isofit}} on the isotopic source data stored in GNIPDataDE (see
example).
}
\examples{

data(GermanFit)
GermanFit
plot(GermanFit)

## The following example takes a lot of time and will therefore not
## be run unless you type: example(GermanFit, run.dontrun=TRUE)

\dontrun{
## We load the data for Germany
data(GNIPDataDE)

## We prepare the data using the function \\code{\\link{queryGNIP}}
GNIPDataDEagg <- queryGNIP(data = GNIPDataDE)

## We fit the isoscape model
GermanFit <- isofit(iso.data = GNIPDataDEagg,
                    mean.model.fix = list(elev = TRUE, lat.abs = TRUE),
                    mean.model.rand = list("uncorr" = TRUE),
                    disp.model.rand = list("uncorr" = TRUE)
                    )

## We created the object GermanFit stored in this package using
## save(GermanFit, file = "GermanFit.rda", compress = "xz")

GermanFit
plot(GermanFit)
}


}
\seealso{
\code{\link{isofit}} for information about how to fit an isoscape
model
}
\keyword{datasets}
\keyword{models}
