% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getelev.R
\name{getprecip}
\alias{getprecip}
\alias{downloadfile}
\title{Download an rasters of monthly precipitation from internet}
\source{
\url{http://worldclim.org/version2}
}
\usage{
getprecip(path = NULL, overwrite = FALSE, verbose = interactive())
}
\arguments{
\item{path}{A \var{string} indicating where to store the file on the hard
drive}

\item{overwrite}{A \var{logical} indicating if an existing file should be
re-downloaded}

\item{verbose}{A \var{logical} indicating whether information about the
progress of the procedure should be displayed or not while the function is
running. By default verbose is \code{TRUE} if users use an interactive R
session and \code{FALSE} otherwise. If a \var{numeric} is provided instead,
additional information about the download will be provided if the number is
greater than 1.}
}
\description{
The function \code{getprecip} allows for the download of rasters of monthly
precipitation from internet. It downloads the "precipitation (mm) WorldClim
Version2" at a spatial resolution of 30 seconds (~1 km2). After download, the
function also unzip the file. The function
\code{getprecip} uses the generic function \code{downloadfile} that can also be
used to download directly other files.
}
\details{
In the argument "path" is not provided, the file will be stored in the 
current working directory. The functions can create new directories, so you 
can also indicate a new path. The integrity of the elevation raster is tested
after a call to \code{getprecip}. In case of corruption, try downloading the
file again, specifying overwrite = TRUE to overwrite the corrupted file.
}
\examples{

## To download the monthly precipitation
## in your current working
## directory, just type:
## getprecip()
## Mind that the file weights ca. 1GB!
}
