% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4classes.R
\docType{class}
\name{ISOFIND-class}
\alias{ISOFIND-class}
\title{Class ISOFIND}
\description{
Class ISOFIND
}
\section{Slots}{

\describe{
\item{\code{sample}}{a list of SpatRaster objects storing the assignment info for each sample}

\item{\code{group}}{a SpatRaster storing the group assignment info}

\item{\code{sp_points}}{a list of SpatVector storing the spatial points for sources, calibration and assignment samples}
}}

