\name{Deprecated functions}
\alias{Calibfit}
\alias{GetElev}
\alias{Isofit}
\alias{Isorix}
\alias{Isoscape}
\alias{Isosim}
\alias{QueryGNIP}
\alias{RElevate}

\title{Deprecated functions}

\description{The function you asked help for has been deprecated (i.e. it does not longer exists). A new function with a different name is surely doing the old job.}

\examples{
## To see all current function names, try:
ls("package:IsoriX")
}