% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4classes.R
\docType{class}
\name{ISOSCAPE-class}
\alias{ISOSCAPE-class}
\title{Class ISOSCAPE}
\description{
Class ISOSCAPE
}
\section{Slots}{

\describe{
\item{\code{isoscapes}}{a SpatRaster storing the isoscapes}

\item{\code{sp_points}}{a list of spatial points}
}}

