% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load data}
\usage{
load_data(
  data,
  data_needed = list(pre = FALSE, pre_cv = FALSE, rd0 = FALSE, tmp = FALSE, dtr =
    FALSE, reh = FALSE, tmn = FALSE, tmx = FALSE, sunp = FALSE, frs = FALSE, wnd = FALSE,
    elv = FALSE),
  long_min,
  long_max,
  lat_min,
  lat_max,
  split_by = NULL
)
}
\arguments{
\item{data}{A \var{dataframe} containing raw isotopic measurements of sources}

\item{data_needed}{A \var{list} of \var{logical} indicating which variable to add.
The following data are available:
\describe{
\item{pre}{precipitation (millimetres/month)}
\describe{
\item{pre_cv}{cv of precipitation (percent)}
}
\item{rd0}{wet-days (number days with >0.1mm rain per month)}
\item{tmp}{mean temperature (degrees Celsius)}
\item{tmn}{minimum temperature (degrees Celsius)}
\item{tmx}{maximum temperature (degrees Celsius)}
\item{dtr}{mean diurnal temperature range (degrees Celsius)}
\item{reh}{relative humidity (percent)}
\item{sunp}{sunshine (percent of maximum possible (percent of day length))}
\item{frs}{ground-frost (number of days with ground-frost per month)}
\item{wnd}{10 metre windspeed (metres/second)}
\item{elv}{elevation (automatically converted to metres)}
}}

\item{long_min}{A \var{numeric} indicating the minimum longitude to select
from. Should be a number between -180 and 180. If not provided, -180 will
be considered.}

\item{long_max}{A \var{numeric} indicating the maximal longitude to select
from. Should be a number between -180 and 180. If not provided, 180 will be
considered.}

\item{lat_min}{A \var{numeric} indicating the minimum latitude to select
from. Should be a number between -90 and 90. If not provided, -90 will be
considered.}

\item{lat_max}{A \var{numeric} indicating the maximal latitude to select
from. Should be a number between -90 and 90. If not provided, 90 will be
considered.}

\item{split_by}{A \var{string} indicating whether data should be aggregated
per location (\code{split_by = NULL}, the default), per location:month
combination (\code{split_by = "month"}), or per location:year combination
(\code{split_by = "year"}).}
}
\value{
This function returns a \var{list} of two elements: A \var{dataframe}
containing the extended data and a \var{list} containing the raster.
}
\description{
This function downloads and adds climatic data to improve the fit.
}
\details{
Data comes from the University of East Anglia Climate Research Unit (CRU)
climatology data (see \url{\link{https://docs.ropensci.org/getCRUCLdata/}}).
}
\examples{
load_data(data = GNIPDataDE, 
          data_needed = list(pre = TRUE, tmp = TRUE, elv = FALSE), split_by = "year")  
}
