% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsoriX-datasets.R
\docType{data}
\name{OceanMask_terra}
\alias{OceanMask_terra}
\title{Mask of world oceans (terra format)}
\format{
A \emph{SpatVector} object
}
\source{
This \emph{SpatVector} is derived from the
\link{CountryBorders_terra}. See example for details on how we created the
dataset.
}
\description{
This dataset contains a polygon shapefile that can be used to mask large
bodies of water.
}
\examples{

plot(OceanMask_terra, col='blue')

## How did we create this file?

## Uncomment the following to create the file as we did
# if (require(terra)) {
#    worldlimit <- as.polygons(ext(CountryBorders_terra))
#    crs(worldlimit) <- crs(CountryBorders_terra)
#    OceanMask_terra <- worldlimit - CountryBorders_terra
# }


}
\seealso{
\itemize{
\item \link{OceanMask} for the same file stored as a \pkg{sp} object
\item \link{CountryBorders_terra} for another polygon shapefile used to embellish the plots
}
}
\keyword{datasets}
