% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsoriX-datasets.R
\docType{data}
\name{AssignDataBat}
\alias{AssignDataBat}
\title{Assignment dataset for bat species}
\format{
A *dataframe* with 14 observations on 4 variables:
\tabular{rlll}{
[, 1] \tab sample_ID \tab (*factor*) \tab Identification of the animal\cr
[, 2] \tab lat \tab (*numeric*) \tab Latitude coordinate (decimal degrees)\cr
[, 2] \tab lomg \tab (*numeric*) \tab Longitude coordinate (decimal degrees)\cr
[, 4] \tab sample_value \tab (*numeric*) \tab Hydrogen delta value of the tissue\cr }
}
\source{
data directly provided by the authors of the following publication
}
\description{
This dataset contains data from Voigt & Lenhert (2019). It contains hydrogen
delta values of fur keratin from common noctule bats (\emph{Nyctalus noctula})
killed at wind turbines in northern Germany. The data can be used as an
example to perform assignments using the function [isofind].
}
\examples{

head(AssignDataBat)
str(AssignDataBat)

}
\references{
Voigt CC & Lehnert L (2019). Tracking of movements of terrestrial
mammals using stable isotopes. In Hobson KA, Wassenaar LI (eds.), Tracking Animal
Migration with Stable Isotopes, second edition. Academic Press, London.
}
\seealso{
[isofind] to perform assignments
}
\keyword{datasets}
