% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsoriX-dataset.R
\docType{data}
\name{CountryBorders}
\alias{CountryBorders}
\title{Borders of world CountryBorders}
\format{A \var{SpatialPolygons}}
\source{
This \var{SpatialPolygons} is derived from the
\code{\link[maps]{world}} of the package \pkg{maps}. Please refer to this
other package for description and sources of this dataset. See example for
details on how we created the dataset.
}
\description{
This dataset contains a polygon shapefile that can be used to plot the
borders of CountryBorders.
}
\examples{

if(require(sp))
  plot(CountryBorders, border="red", col="darkgrey")


### HOW DID WE CREATE THIS FILE?
### (This example takes some time and will therefore not be run
###  unless you type: example(CountryBorders, run.dontrun = TRUE) )

\dontrun{
if(require(maps) & require(maptools) & require(raster) & require(rgeos)){
    worldmap <- map("world", fill = TRUE, plot = FALSE)
    CountryBorders <- map2SpatialPolygons(worldmap, IDs = worldmap$names)
    CountryBorders <- gBuffer(CountryBorders, byid = TRUE, width = 0)
    proj4string(CountryBorders) <- CRS("+ proj = longlat + datum = WGS84")
    CountryBorders
    ## uncomment the following to store the file:
    #save(CountryBorders, file = "CountryBorders.rda", compress = "xz")
}
}


}
\seealso{
\code{\link{OceanMask}} for another polygon shapefile used to
embellish the plots
}
\keyword{datasets}
