% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibfit.R, R/getelev.R, R/isofind.R,
%   R/isofit.R, R/isoscape.R, R/isosim.R, R/prepareData.R, R/relevate.R
\name{Calibfit}
\alias{Calibfit}
\alias{GetElev}
\alias{IsoriX-defunct}
\alias{Calibfit}
\alias{GetElev}
\alias{Isofit}
\alias{Isorix}
\alias{Isoscape}
\alias{Isosim}
\alias{QueryGNIP}
\alias{queryGNIP}
\alias{RElevate}
\alias{Isorix}
\alias{Isofit}
\alias{Isoscape}
\alias{Isosim}
\alias{queryGNIP}
\alias{QueryGNIP}
\alias{RElevate}
\title{Deprecated functions}
\usage{
Calibfit(...)

GetElev(...)

Isorix(...)

Isofit(...)

Isoscape(...)

Isosim(...)

queryGNIP(...)

QueryGNIP(...)

RElevate(...)
}
\arguments{
\item{...}{The call of the deprecated function}
}
\description{
The function you asked help for has been deprecated (i.e. it does not longer
exists). A new function with a different name is surely doing the old job.
}
