% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{IsoriX-defunct}
\alias{IsoriX-defunct}
\alias{Calibfit}
\alias{GetElev}
\alias{Isofit}
\alias{Isorix}
\alias{Isoscape}
\alias{Isosim}
\alias{QueryGNIP}
\alias{queryGNIP}
\alias{prepdata}
\alias{prepelev}
\alias{prepiso}
\alias{RElevate}
\alias{relevate}
\title{Defunct and deprecated functions}
\usage{
Calibfit(...)

GetElev(...)

Isofit(...)

Isorix(...)

Isoscape(...)

Isosim(...)

QueryGNIP(...)

queryGNIP(...)

prepdata(...)

prepelev(...)

prepiso(...)

RElevate(...)

relevate(...)
}
\arguments{
\item{...}{The call of the defunct or deprecated function}
}
\description{
The function you asked help for has been defunct (i.e. it does not longer
exists) or deprecated (i.e. it will disappear soon). A new function with a
different name is surely doing the old job.
}
