% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsoriX-datasets.R
\docType{data}
\name{OceanMask}
\alias{OceanMask}
\title{Mask of world oceans (sp format)}
\format{
A \emph{SpatialPolygons} object
}
\source{
This \emph{SpatialPolygons} is derived from the
\link{CountryBorders_terra}. See example for details on how we created the
dataset.
}
\description{
This dataset contains a polygon shapefile that can be used to mask large
bodies of water.
}
\examples{

if (require(sp)) {
   plot(OceanMask, col='blue')
}

## How did we create this file?

## Uncomment the following to create the file as we did
# if (require(terra)) {
#    worldlimit <- as.polygons(ext(CountryBorders_terra))
#    crs(worldlimit) <- crs(CountryBorders_terra)
#    OceanMask_terra <- worldlimit - CountryBorders_terra
#    OceanMask <- as(OceanMask_terra, "Spatial")
# }


}
\seealso{
\itemize{
\item \link{OceanMask_terra} for the same file stored as a \pkg{terra} object
\item \link{CountryBorders} for another polygon shapefile used to embellish the plots
}
}
\keyword{datasets}
