% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsoriX-dataset.R
\docType{data}
\name{oceanmask}
\alias{oceanmask}
\title{Mask of world oceans}
\format{A \var{SpatialPolygons} object}
\source{
This \var{SpatialPolygons} is derived from the
\code{\link{countries}}. See example for details on how we created the
dataset.
}
\description{
This dataset contains a polygon shapefile that can be used to mask large
bodies of water.
}
\examples{

data(oceanmask)
if(require(sp))
    plot(oceanmask, col='blue')

## HOW DID WE CREATE THIS FILE?
## (This example takes some time and will therefore not be run
##  unless you type: example(oceanmask, run.dontrun=TRUE) )
  
\dontrun{
if(require(raster) & require(rgeos)){
    worldlimit <- as(extent(countries), "SpatialPolygons")
    proj4string(worldlimit) <- crs(countries)
    oceanmask <- gDifference(worldlimit, countries)  
    oceanmask
    ## uncomment the following to store the file:
    #save(oceanmask, file="oceanmask.rda", compress="xz")
}
}


}
\seealso{
\code{\link{countries}} for another polygon shapefile used to
embellish the plots
}
\keyword{datasets}
