% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsoriX-datasets.R
\docType{data}
\name{OceanMask}
\alias{OceanMask}
\title{Mask of world oceans}
\format{
A \emph{SpatVector} object
}
\source{
See example for details on how we created the dataset.
}
\description{
This dataset contains a polygon SpatVector (from \pkg{terra}).
It can be used to mask large bodies of water.
}
\examples{

plot(OceanMask, col='blue')

## How did we create this file?

## Uncomment the following to create the file as we did
# if (require(terra)) {
#   worldlimit <- vect(ext(CountryBorders))
#   crs(worldlimit) <- crs(CountryBorders)
#   OceanMask <- worldlimit - CountryBorders
#   #saveRDS(OceanMask, file = "IsoriX/inst/extdata/OceanMask.rds", compress = "xz")
# }


}
\seealso{
\itemize{
\item \link{CountryBorders} for another polygon used to embellish the plots
}
}
\keyword{datasets}
