% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsoriX-dataset.R
\docType{data}
\name{OceanMask}
\alias{OceanMask}
\title{Mask of world oceans}
\format{A \var{SpatialPolygons} object}
\source{
This \var{SpatialPolygons} is derived from the
\code{\link{CountryBorders}}. See example for details on how we created the
dataset.
}
\description{
This dataset contains a polygon shapefile that can be used to mask large
bodies of water.
}
\examples{

if(require(sp)) {
  plot(OceanMask, col='blue')
}

## How did we create this file?

if(require(raster) & require(rgeos)){
    worldlimit <- as(extent(CountryBorders), "SpatialPolygons")
    proj4string(worldlimit) <- crs(CountryBorders)
    OceanMask <- gDifference(worldlimit, CountryBorders)  
    OceanMask
    
## Uncomment the following to store the file as we did
#save(OceanMask, file = "OceanMask.rda", compress = "xz")

}


}
\seealso{
\code{\link{CountryBorders}} for another polygon shapefile used to
embellish the plots
}
\keyword{datasets}
