% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{options}
\alias{options}
\alias{IsoriX.options}
\alias{IsoriX.getOption}
\alias{IsoriX.getOption}
\title{Setting and displaying the options of the package}
\usage{
IsoriX.options(...)

IsoriX.getOption(x)
}
\arguments{
\item{...}{A named value or a list of named values. The following values, with their defaults, are used:
\describe{
  \item{example_maxtime}{The number of seconds allowed for a given example to run. It is used to control whether the longer examples should be run or not based on the comparison between this option and the approximate running time of the example on our computers.}
  \item{Ncpu}{An \var{integer} corresponding to the number of cores to be used (for future version)}
}}

\item{x}{A character string holding an option name.}
}
\value{
The options invisibly in an object called \code{.IsoriX.data$options}
}
\description{
Setting and displaying the options of the package
}
\examples{
IsoriX.options()
IsoriX.getOption("example_maxtime")

\dontrun{
IsoriX.options(example_maxtime = 30)
}
}
