% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{sp.polygons}
\alias{sp.points}
\alias{CRS}
\alias{raster}
\alias{values}
\alias{area}
\alias{extent}
\alias{extent<-}
\alias{projection<-}
\alias{shift}
\alias{RdBuTheme}
\alias{levelplot}
\alias{grid.text}
\alias{xyplot}
\alias{panel.points}
\alias{layer}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{grid}{\code{\link[grid]{grid.text}}}

  \item{lattice}{\code{\link[lattice]{xyplot}}, \code{\link[lattice]{panel.points}}}

  \item{latticeExtra}{\code{\link[latticeExtra]{layer}}}

  \item{raster}{\code{\link[raster]{raster}}, \code{\link[raster]{values}}, \code{\link[raster]{area}}, \code{\link[raster]{extent}}, \code{\link[raster]{extent<-}}, \code{\link[raster]{projection<-}}, \code{\link[raster]{shift}}}

  \item{rasterVis}{\code{\link[rasterVis]{RdBuTheme}}, \code{\link[rasterVis]{levelplot}}}

  \item{sp}{\code{\link[sp]{sp.polygons}}, \code{\link[sp]{sp.points}}, \code{\link[sp]{CRS}}}
}}

