% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{gpar}
\alias{grid.text}
\alias{panel.points}
\alias{xyplot}
\alias{layer}
\alias{levelplot}
\alias{RdBuTheme}
\alias{get_ranPars}
\alias{plot}
\alias{points}
\alias{polys}
\alias{rast}
\alias{extract}
\alias{crs}
\alias{crs<-}
\alias{ext}
\alias{ext<-}
\alias{shift}
\alias{values}
\alias{cellSize}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{grid}{\code{\link[grid]{gpar}}, \code{\link[grid]{grid.text}}}

  \item{lattice}{\code{\link[lattice:llines]{panel.points}}, \code{\link[lattice]{xyplot}}}

  \item{latticeExtra}{\code{\link[latticeExtra]{layer}}}

  \item{rasterVis}{\code{\link[rasterVis:levelplot-methods]{levelplot}}, \code{\link[rasterVis:rasterTheme]{RdBuTheme}}}

  \item{spaMM}{\code{\link[spaMM:get_fittedPars]{get_ranPars}}}

  \item{terra}{\code{\link[terra]{cellSize}}, \code{\link[terra]{crs}}, \code{\link[terra:crs]{crs<-}}, \code{\link[terra]{ext}}, \code{\link[terra:ext]{ext<-}}, \code{\link[terra]{extract}}, \code{\link[terra]{plot}}, \code{\link[terra:lines]{points}}, \code{\link[terra:lines]{polys}}, \code{\link[terra]{rast}}, \code{\link[terra]{shift}}, \code{\link[terra]{values}}}
}}

